getting assignment of dicts/lists tolerably typed
