+++ pyproject.toml
@@ -11,7 +11,7 @@ packages = [
 [tool.poetry.dependencies]
 python = "^3.8"
 bitarray = "^2.6.0"
-crc = "^1.3.0"
+crc = "^2.0.0"
 
 [tool.poetry.dev-dependencies]
 black = "^22.1"
