+++ src/easyprotocol/base/parse_field.py
@@ -2,40 +2,45 @@
 # from __future__ import annotations
 
 # from collections import OrderedDict
-# from typing import Any, Generic, Literal, TypeVar, Union, overload
+# from typing import Any, Generic, Literal, Sequence, TypeVar, Union, cast, overload
 
 # from bitarray import bitarray
 # from typing_extensions import override
 
+# from easyprotocol.base.parse_dict_generic import ParseGenericDict
 # from easyprotocol.base.parse_generic import (
 #     DEFAULT_ENDIANNESS,
 #     UNDEFINED,
 #     ParseGeneric,
 #     endianT,
 # )
+# from easyprotocol.base.parse_list_generic import ParseListGeneric
+# from easyprotocol.base.parse_generic_value import ParseGenericValue
 # from easyprotocol.base.utils import dataT
 
 # T = TypeVar("T", bound=ParseGeneric[Any])
+# # ParseGenericUnion = TypeVar(
+# #     "ParseGenericUnion", ParseGenericValue[Any], ParseGenericDict[Any], ParseListGeneric[Any], covariant=True
+# # )
+# ParseGenericUnion = Union[ParseGenericValue[T], ParseGenericDict[T], ParseListGeneric[T]]
 
 
 # class ParseFieldGeneric(
-#     ParseGeneric[T],
-#     Generic[T],
+#     ParseGeneric[ParseGenericUnion[T]],
 # ):
 #     """The base parsing object for handling parsing in a convenient (to modify) package."""
 
 #     def __init__(
 #         self,
 #         name: str,
-#         value: T | None = None,
-#         data: dataT = None,
 #         bit_count: int = -1,
+#         data: dataT = None,
 #         string_format: str | None = None,
 #         endian: endianT = DEFAULT_ENDIANNESS,
 #         parent: ParseGeneric[Any] | None = None,
 #         children: OrderedDict[str, ParseGeneric[Any]]
 #         | dict[str, ParseGeneric[Any]]
-#         | list[ParseGeneric[Any]]
+#         | Sequence[ParseGeneric[Any]]
 #         | None = None,
 #     ) -> None:
 #         """Create the base parsing object for handling parsing in a convenient package.
@@ -50,7 +55,6 @@
 #         """
 #         super().__init__(
 #             name=name,
-#             value=value,
 #             data=data,
 #             bit_count=bit_count,
 #             string_format=string_format,
@@ -81,44 +85,25 @@
 #         """
 #         raise NotImplementedError()
 
-#     def get_value(self) -> T:
-#         raise NotImplementedError()
+#     # def get_value(self) -> T:
+#     #     raise NotImplementedError()
 
-#     def set_value(self, value: T) -> None:
-#         raise NotImplementedError()
+#     # def set_value(self, value: T) -> None:
+#     #     raise NotImplementedError()
 
-#     def get_string_value(self) -> str:
-#         """Get a formatted value for the field (for any custom formatting).
+#     # def get_string_value(self) -> str:
+#     #     """Get a formatted value for the field (for any custom formatting).
 
-#         Returns:
-#             the value of the field with custom formatting
-#         """
-#         if self.value is None:
-#             return UNDEFINED
-#         else:
-#             return self._string_format.format(self.value)
+#     #     Returns:
+#     #         the value of the field with custom formatting
+#     #     """
+#     #     if self.value is None:
+#     #         return UNDEFINED
+#     #     else:
+#     #         return self._string_format.format(self.value)
 
 #     def get_bytes_value(self) -> bytes:
 #         return self.__bytes__()
-
-#     def _get_children_generic(self) -> OrderedDict[str, ParseGeneric[T]]:
-#         return self._children
-
-#     def _set_children_generic(
-#         self,
-#         children: OrderedDict[str, ParseGeneric[T]] | dict[str, ParseGeneric[T]] | list[ParseGeneric[T]] | None,
-#     ) -> None:
-#         self._children.clear()
-#         if isinstance(children, (dict, OrderedDict)):
-#             keys = list(children.keys())
-#             for key in keys:
-#                 value = children[key]
-#                 self._children[key] = value
-#                 value.parent = self
-#         elif isinstance(children, list):
-#             for value in children:
-#                 self._children[value._name] = value
-#                 value.parent = self
 
 #     def get_bits(self) -> bitarray:
 #         return self._bits
@@ -139,18 +124,24 @@
 #     def name(self, name: str) -> None:
 #         self._name = name
 
-#     @property
-#     def value(self) -> T:
-#         """Get the parsed value of the field.
+#     def get_parent(self) -> ParseGenericUnion[Any] | None:
+#         return cast(ParseGenericUnion[Any], self._parent)
 
-#         Returns:
-#             the value of the field
-#         """
-#         return self.get_value()
+#     def set_parent(self, parent: ParseGenericUnion[T] | None) -> None:
+#         self._parent = parent
 
-#     @value.setter
-#     def value(self, value: T) -> None:
-#         self.set_value(value)
+#     # @property
+#     # def value(self) -> T:
+#     #     """Get the parsed value of the field.
+
+#     #     Returns:
+#     #         the value of the field
+#     #     """
+#     #     return self.get_value()
+
+#     # @value.setter
+#     # def value(self, value: T) -> None:
+#     #     self.set_value(value)
 
 #     @property
 #     def bits(self) -> bitarray:
@@ -166,17 +157,17 @@
 #         self.set_bits(bits)
 
 #     @property
-#     def parent(self) -> ParseGeneric[T] | None:
+#     def parent(self) -> ParseGenericUnion[T] | None:
 #         """Get the parent of the field.
 
 #         Returns:
 #             the parent of the field
 #         """
-#         return self._parent
+#         return self.get_parent()
 
 #     @parent.setter
-#     def parent(self, parent: ParseGeneric[T] | None) -> None:
-#         self._parent = parent
+#     def parent(self, parent: ParseGenericUnion[T] | None) -> None:
+#         self.set_parent(parent)
 
 #     @property
 #     def string_format(self) -> str:
@@ -190,22 +181,6 @@
 #     @string_format.setter
 #     def string_format(self, fmt: str) -> None:
 #         self._string_format = fmt
-
-#     @property
-#     def children(self) -> OrderedDict[str, ParseGeneric[T]]:
-#         """Get the parse objects that are contained by this one.
-
-#         Returns:
-#             the parse objects that are contained by this one
-#         """
-#         return self._get_children_generic()
-
-#     @children.setter
-#     def children(
-#         self,
-#         children: OrderedDict[str, ParseGeneric[T]] | dict[str, ParseGeneric[T]] | list[ParseGeneric[T]] | None,
-#     ) -> None:
-#         self._set_children_generic(children=children)
 
 #     @property
 #     def endian(self) -> Literal["little", "big"]:
@@ -259,25 +234,25 @@
 #         return f"<{self.__class__.__name__}> {self.__str__()}"
 
 
-# class ParseField(ParseFieldGeneric[Any]):
-#     def __init__(
-#         self,
-#         name: str,
-#         value: Any | None = None,
-#         data: dataT = None,
-#         bit_count: int = -1,
-#         string_format: str | None = None,
-#         endian: endianT = DEFAULT_ENDIANNESS,
-#         parent: ParseFieldGeneric[Any] | None = None,
-#         children: OrderedDict[str, ParseGeneric[T]] | dict[str, ParseGeneric[T]] | list[ParseGeneric[T]] | None = None,
-#     ) -> None:
-#         super().__init__(
-#             name=name,
-#             value=value,
-#             data=data,
-#             bit_count=bit_count,
-#             string_format=string_format,
-#             endian=endian,
-#             parent=parent,
-#             children=children,
-#         )
+# # class ParseField(ParseFieldGeneric[Any]):
+# #     def __init__(
+# #         self,
+# #         name: str,
+# #         value: Any | None = None,
+# #         data: dataT = None,
+# #         bit_count: int = -1,
+# #         string_format: str | None = None,
+# #         endian: endianT = DEFAULT_ENDIANNESS,
+# #         parent: ParseFieldGeneric[Any] | None = None,
+# #         children: OrderedDict[str, ParseGeneric[T]] | dict[str, ParseGeneric[T]] | Sequence[ParseGeneric[T]] | None = None,
+# #     ) -> None:
+# #         super().__init__(
+# #             name=name,
+# #             value=value,
+# #             data=data,
+# #             bit_count=bit_count,
+# #             string_format=string_format,
+# #             endian=endian,
+# #             parent=parent,
+# #             children=children,
+# #         )
