+++ src/easyprotocol/base/parse_generic_list.py
@@ -6,42 +6,39 @@ from typing import (
     Any,
     Generic,
     Iterable,
-    Literal,
+    Iterator,
     MutableSequence,
+    Sequence,
     SupportsIndex,
     TypeVar,
-    Union,
-    cast,
     overload,
 )
 
 from bitarray import bitarray
 
 from easyprotocol.base.parse_generic import DEFAULT_ENDIANNESS, ParseGeneric, endianT
+
+# from easyprotocol.base.parse_generic_value import _T
 from easyprotocol.base.utils import dataT, input_to_bytes
 
-T = TypeVar("T", bound=Any)
+_T = TypeVar("_T")
 
 
-class ParseListGeneric(
-    ParseGeneric[T],
-    MutableSequence[ParseGeneric[T]],
+class ParseGenericList(
+    ParseGeneric[_T],
+    MutableSequence[ParseGeneric[_T]],
+    Generic[_T],
 ):
     """The base parsing object for handling parsing in a convenient package."""
 
     def __init__(
         self,
         name: str,
-        value: list[ParseGeneric[T]] | None = None,
+        default: Sequence[ParseGeneric[_T]] | OrderedDict[str, ParseGeneric[_T]] = list(),
         data: dataT = None,
         bit_count: int = -1,
         string_format: str = "{}",
         endian: endianT = DEFAULT_ENDIANNESS,
-        parent: ParseGeneric[Any] | None = None,
-        children: OrderedDict[str, ParseGeneric[Any]]
-        | dict[str, ParseGeneric[Any]]
-        | list[ParseGeneric[Any]]
-        | None = None,
     ) -> None:
         """Create the base parsing object for handling parsing in a convenient package.
 
@@ -56,18 +53,14 @@ class ParseListGeneric(
             bit_count=bit_count,
             string_format=string_format,
             endian=endian,
-            parent=parent,
         )
 
-        if children is not None:
-            if isinstance(children, dict):
-                self._set_children_generic(children)
-            else:
-                self._set_children_generic(OrderedDict({val._name: val for val in children}))
+        if isinstance(default, dict):
+            self._set_children_generic(default)
+        else:
+            self._set_children_generic(OrderedDict({val._name: val for val in default}))
         if data is not None:
             self.parse(data)
-        elif value is not None:
-            self.set_value(value)
 
     def parse(self, data: dataT) -> bitarray:
         """Parse bytes that make of this protocol field into meaningful data.
@@ -83,8 +76,8 @@ class ParseListGeneric(
             bit_data = field.parse(data=bit_data)
         return bit_data
 
-    def insert(self, index: int | slice, value: ParseGeneric[T] | list[ParseGeneric[T]]) -> None:
-        c: OrderedDict[str, ParseGeneric[T]] = OrderedDict()
+    def insert(self, index: int | slice, value: ParseGeneric[_T] | Sequence[ParseGeneric[_T]]) -> None:
+        c: OrderedDict[str, ParseGeneric[_T]] = OrderedDict()
         for i, v in enumerate(self._children.values()):
             if index == i:
                 if isinstance(value, ParseGeneric):
@@ -95,10 +88,10 @@ class ParseListGeneric(
             c[v._name] = self._children[v._name]
         self._children = c
 
-    def append(self, value: ParseGeneric[T]) -> None:
+    def append(self, value: ParseGeneric[_T]) -> None:
         self._children[value.name] = value
 
-    def get_value(self) -> list[ParseGeneric[T]]:
+    def get_value(self) -> Sequence[ParseGeneric[_T]]:
         """Get the parsed value of the field.
 
         Returns:
@@ -108,7 +101,7 @@ class ParseListGeneric(
 
     def set_value(
         self,
-        value: list[ParseGeneric[T]] | dict[str, ParseGeneric[T]] | OrderedDict[str, ParseGeneric[T]],
+        value: Sequence[ParseGeneric[_T]] | OrderedDict[str, ParseGeneric[_T]],
     ) -> None:
         if isinstance(value, (dict, OrderedDict)):
             values = list(value.values())
@@ -170,7 +163,7 @@ class ParseListGeneric(
         return f"<{self.__class__.__name__}> {self.__str__()}"
 
     @property
-    def value(self) -> list[ParseGeneric[T]]:
+    def value(self) -> Sequence[ParseGeneric[_T]]:
         """Get the parsed value of the field.
 
         Returns:
@@ -179,18 +172,18 @@ class ParseListGeneric(
         return self.get_value()
 
     @value.setter
-    def value(self, value: list[ParseGeneric[T]]) -> None:
+    def value(self, value: Sequence[ParseGeneric[_T]]) -> None:
         self.set_value(value=value)
 
     @overload
-    def __getitem__(self, index: int) -> ParseGeneric[T]:
+    def __getitem__(self, index: int) -> ParseGeneric[_T]:
         ...
 
     @overload
-    def __getitem__(self, index: slice) -> list[ParseGeneric[T]]:
+    def __getitem__(self, index: slice) -> Sequence[ParseGeneric[_T]]:
         ...
 
-    def __getitem__(self, index: int | slice) -> ParseGeneric[T] | list[ParseGeneric[T]]:
+    def __getitem__(self, index: int | slice) -> ParseGeneric[_T] | Sequence[ParseGeneric[_T]]:
         vs = list(self._children.values())[index]
         if isinstance(vs, list):
             return [v for v in vs]
@@ -208,14 +201,14 @@ class ParseListGeneric(
             self._children = OrderedDict({k: v for k, v in self._children.items() if k != item.name})
 
     @overload
-    def __setitem__(self, index: SupportsIndex, value: ParseGeneric[T]) -> None:
+    def __setitem__(self, index: SupportsIndex, value: ParseGeneric[_T]) -> None:
         ...
 
     @overload
-    def __setitem__(self, index: slice, value: Iterable[ParseGeneric[T]]) -> None:
+    def __setitem__(self, index: slice, value: Iterable[ParseGeneric[_T]]) -> None:
         ...
 
-    def __setitem__(self, index: SupportsIndex | slice, value: ParseGeneric[T] | Iterable[ParseGeneric[T]]) -> None:
+    def __setitem__(self, index: SupportsIndex | slice, value: ParseGeneric[_T] | Iterable[ParseGeneric[_T]]) -> None:
         indexed_keys = list(self._children.keys())[index]
         c: OrderedDict[str, ParseGeneric[Any]] = OrderedDict()
         for existing_key in self._children:
@@ -238,12 +231,12 @@ class ParseListGeneric(
     def __len__(self) -> int:
         return len(self._children)
 
-    def _get_children_generic(self) -> OrderedDict[str, ParseGeneric[T]]:
+    def _get_children_generic(self) -> OrderedDict[str, ParseGeneric[_T]]:
         return self._children
 
     def _set_children_generic(
         self,
-        children: OrderedDict[str, ParseGeneric[T]] | dict[str, ParseGeneric[T]] | list[ParseGeneric[T]] | None,
+        children: OrderedDict[str, ParseGeneric[_T]] | Sequence[ParseGeneric[_T]] | None,
     ) -> None:
         self._children.clear()
         if isinstance(children, (dict, OrderedDict)):
@@ -269,3 +262,6 @@ class ParseListGeneric(
     @value.setter
     def value(self, value: ParseGeneric[Any] | None) -> None:
         self._set_parent_generic(value)
+
+    def __iter__(self) -> Iterator[ParseGeneric[Any]]:
+        return self._children.values().__iter__()
