+++ tests/test_parse_enum.py
@@ -1,22 +1,25 @@
 import struct
 from collections import OrderedDict
-from enum import IntFlag
+from enum import IntEnum
 from typing import Any
 
 import pytest
 from bitarray import bitarray
 from parse_data import ParseData
-from test_uint import check_int_properties, check_int_value
+from test_parse_uint import check_int_properties, check_int_value
 
 from easyprotocol.base.parse_generic import DEFAULT_ENDIANNESS
-from easyprotocol.fields.flags import FlagsField
+from easyprotocol.fields.enum import EnumField
 
 
-def check_flags_strings(
-    obj: FlagsField[IntFlag],
+def check_enum_strings(
+    obj: EnumField[IntEnum],
     tst: ParseData,
 ) -> None:
-
+    assert tst.string_format.format(tst.value.name) == obj.string, (
+        f"{obj}: obj.string is not the expected value "
+        + f"({tst.string_format.format(tst.value)} != expected value: {obj.string})"
+    )
     assert len(obj.string) > 0, f"{obj}: obj.string is not the expected value " + f"(? != expected value: {obj.string})"
     assert tst.name in str(obj), f"{obj}: obj.name is not in the object's string vale ({obj.name} not in {str(obj)})"
     assert obj.string in str(
@@ -31,8 +34,8 @@ def check_flags_strings(
     ), f"{obj}: obj.__class__.__name__ is not in the object's repr vale ({obj.__class__.__name__} not in {repr(obj)})"
 
 
-def check_flags(
-    obj: FlagsField[Any],
+def check_enum(
+    obj: EnumField[Any],
     tst: ParseData,
 ) -> None:
     check_int_value(
@@ -43,24 +46,27 @@ def check_flags(
         obj=obj,
         tst=tst,
     )
-    check_flags_strings(
+    check_enum_strings(
         obj=obj,
         tst=tst,
     )
 
 
-class TestingFlags(IntFlag):
-    NONE = 0
-    ONE = 1
-    TWO = 2
-    FOUR = 4
-    EIGHT = 8
+class TestEnumerating(IntEnum):
+    ZERO = 0b000
+    ONE = 0b001
+    TWO = 0b010
+    THREE = 0b011
+    FOUR = 0b100
+    FIVE = 0b101
+    SIX = 0b110
+    SEVEN = 0b111
 
 
-class TestFlags:
-    def test_flags_create_empty(self) -> None:
+class TestEnums:
+    def test_enum_create_empty_big_endian(self) -> None:
+        value = TestEnumerating.ZERO
         bit_count = 2
-        value = TestingFlags.NONE
         byte_data = struct.pack("B", value.value)
         bits_data = bitarray(endian="little")
         bits_data.frombytes(byte_data)
@@ -75,19 +81,19 @@ class TestFlags:
             endian=DEFAULT_ENDIANNESS,
             children=OrderedDict(),
         )
-        obj = FlagsField(
+        obj = EnumField(
             name=tst.name,
             bit_count=bit_count,
-            flags_type=TestingFlags,
-            default=TestingFlags.NONE,
+            enum_type=TestEnumerating,
+            default=TestEnumerating.ZERO,
         )
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
-    def test_flags_create_parse_one(self) -> None:
-        value = TestingFlags.ONE
+    def test_enum_create_empty_little_endian(self) -> None:
+        value = TestEnumerating.ZERO
         bit_count = 2
         byte_data = struct.pack("B", value.value)
         bits_data = bitarray(endian="little")
@@ -103,55 +109,54 @@ class TestFlags:
             endian=DEFAULT_ENDIANNESS,
             children=OrderedDict(),
         )
-        obj = FlagsField(
+        obj = EnumField(
             name=tst.name,
             bit_count=bit_count,
-            flags_type=TestingFlags,
+            enum_type=TestEnumerating,
+            default=TestEnumerating.ZERO,
+        )
+        check_enum(
+            obj=obj,
+            tst=tst,
+        )
+
+    def test_enum_create_parse(self) -> None:
+        value = TestEnumerating.THREE
+        bit_count = 2
+        byte_data = struct.pack("B", value.value)
+        bits_data = bitarray(endian="little")
+        bits_data.frombytes(byte_data)
+        bits_data = bits_data[:bit_count]
+        tst = ParseData(
+            name="test",
+            value=value,
+            string_format="{}",
+            byte_data=byte_data,
+            bits_data=bits_data,
+            parent=None,
+            endian=DEFAULT_ENDIANNESS,
+            children=OrderedDict(),
+        )
+        obj = EnumField(
+            name=tst.name,
+            bit_count=bit_count,
+            enum_type=TestEnumerating,
             data=tst.byte_data,
-            default=TestingFlags.NONE,
+            default=TestEnumerating.ZERO,
         )
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
-    def test_flags_create_parse_multiple(self) -> None:
-        value = TestingFlags.ONE | TestingFlags.TWO | TestingFlags.FOUR | TestingFlags.EIGHT
-        bit_count = 4
-        byte_data = struct.pack("B", value.value)
-        bits_data = bitarray(endian="little")
-        bits_data.frombytes(byte_data)
-        bits_data = bits_data[:bit_count]
-        tst = ParseData(
-            name="test",
-            value=value,
-            string_format="{}",
-            byte_data=byte_data,
-            bits_data=bits_data,
-            parent=None,
-            endian=DEFAULT_ENDIANNESS,
-            children=OrderedDict(),
-        )
-        obj = FlagsField(
-            name=tst.name,
-            bit_count=bit_count,
-            flags_type=TestingFlags,
-            data=tst.byte_data,
-            default=TestingFlags.NONE,
-        )
-        check_flags(
-            obj=obj,
-            tst=tst,
-        )
-
-    def test_flags_create_truncate(self) -> None:
-        value1 = TestingFlags.ONE | TestingFlags.TWO | TestingFlags.FOUR | TestingFlags.EIGHT
-        value2 = TestingFlags.ONE | TestingFlags.TWO
+    def test_enum_create_truncate(self) -> None:
         bit_count = 2
+        value1 = TestEnumerating.SEVEN
         byte_data1 = struct.pack("B", value1.value)
         bits_data1 = bitarray(endian="little")
         bits_data1.frombytes(byte_data1)
         bits_data1 = bits_data1[:bit_count]
+        value2 = TestEnumerating.THREE
         byte_data2 = struct.pack("B", value2.value)
         bits_data2 = bitarray(endian="little")
         bits_data2.frombytes(byte_data2)
@@ -166,29 +171,29 @@ class TestFlags:
             endian=DEFAULT_ENDIANNESS,
             children=OrderedDict(),
         )
-        obj = FlagsField(
+        obj = EnumField(
             name=tst.name,
             bit_count=bit_count,
-            flags_type=TestingFlags,
-            data=byte_data2,
-            default=TestingFlags.NONE,
+            enum_type=TestEnumerating,
+            data=byte_data1,
+            default=TestEnumerating.ZERO,
         )
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
-    def test_flags_create_invalid(self) -> None:
+    def test_enum_create_invalid(self) -> None:
         with pytest.raises(TypeError):
-            FlagsField(
-                name="invalid",
+            EnumField(
+                name="enum",
                 bit_count=4,
-                flags_type=TestingFlags,
-                data="pickles",  # pyright:ignore[reportGeneralTypeIssues]
+                enum_type=TestEnumerating,
+                data=tuple((1, 2, "ff")),  # pyright:ignore[reportGeneralTypeIssues]
             )
 
-    def test_flags_set_name(self) -> None:
-        value = TestingFlags.ONE | TestingFlags.TWO | TestingFlags.FOUR | TestingFlags.EIGHT
+    def test_enum_set_name(self) -> None:
+        value = TestEnumerating.SEVEN
         bit_count = 4
         byte_data = struct.pack("B", value.value)
         bits_data = bitarray(endian="little")
@@ -204,28 +209,28 @@ class TestFlags:
             endian=DEFAULT_ENDIANNESS,
             children=OrderedDict(),
         )
-        obj = FlagsField(
+        obj = EnumField(
             name=tst.name,
             bit_count=bit_count,
-            flags_type=TestingFlags,
+            enum_type=TestEnumerating,
             data=tst.byte_data,
-            default=TestingFlags.NONE,
+            default=TestEnumerating.ZERO,
         )
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
         tst.name = "new_name"
         obj.name = tst.name
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
-    def test_flags_set_value(self) -> None:
-        value1 = TestingFlags.ONE
-        value2 = TestingFlags.ONE | TestingFlags.TWO | TestingFlags.FOUR | TestingFlags.EIGHT
+    def test_enum_set_value(self) -> None:
+        value1 = TestEnumerating.ONE
+        value2 = TestEnumerating.FOUR
         bit_count = 4
         byte_data1 = struct.pack("B", value1.value)
         bits_data1 = bitarray(endian="little")
@@ -245,14 +250,14 @@ class TestFlags:
             endian=DEFAULT_ENDIANNESS,
             children=OrderedDict(),
         )
-        obj = FlagsField(
+        obj = EnumField(
             name=tst.name,
             bit_count=bit_count,
-            flags_type=TestingFlags,
+            enum_type=TestEnumerating,
             data=tst.byte_data,
-            default=TestingFlags.NONE,
+            default=TestEnumerating.ZERO,
         )
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
@@ -261,14 +266,14 @@ class TestFlags:
         tst.value = value2
         tst.byte_data = byte_data2
         tst.bits_data = bits_data2
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
-    def test_flags_set_bits(self) -> None:
-        value1 = TestingFlags.ONE
-        value2 = TestingFlags.ONE | TestingFlags.TWO | TestingFlags.FOUR | TestingFlags.EIGHT
+    def test_enum_set_bits(self) -> None:
+        value1 = TestEnumerating.ONE
+        value2 = TestEnumerating.SIX
         bit_count = 4
         byte_data1 = struct.pack("B", value1.value)
         bits_data1 = bitarray(endian="little")
@@ -288,14 +293,14 @@ class TestFlags:
             endian=DEFAULT_ENDIANNESS,
             children=OrderedDict(),
         )
-        obj = FlagsField(
+        obj = EnumField(
             name=tst.name,
             bit_count=bit_count,
-            flags_type=TestingFlags,
+            enum_type=TestEnumerating,
             data=tst.byte_data,
-            default=TestingFlags.NONE,
+            default=TestEnumerating.ZERO,
         )
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
@@ -304,13 +309,13 @@ class TestFlags:
         tst.value = value2
         tst.byte_data = byte_data2
         tst.bits_data = bits_data2
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
-    def test_flags_set_parent(self) -> None:
-        value = TestingFlags.ONE | TestingFlags.TWO | TestingFlags.FOUR | TestingFlags.EIGHT
+    def test_enum_set_parent(self) -> None:
+        value = TestEnumerating.FIVE
         bit_count = 4
         byte_data = struct.pack("B", value.value)
         bits_data = bitarray(endian="little")
@@ -326,28 +331,27 @@ class TestFlags:
             endian=DEFAULT_ENDIANNESS,
             children=OrderedDict(),
         )
-        obj = FlagsField(
+        obj = EnumField(
             name=tst.name,
             bit_count=bit_count,
-            flags_type=TestingFlags,
+            enum_type=TestEnumerating,
             data=tst.byte_data,
-            default=TestingFlags.NONE,
+            default=TestEnumerating.ZERO,
         )
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
 
-        tst.parent = FlagsField(
+        tst.parent = EnumField(
             name="parent",
             bit_count=bit_count,
-            flags_type=TestingFlags,
+            enum_type=TestEnumerating,
             data=tst.byte_data,
-            default=TestingFlags.NONE,
+            default=TestEnumerating.ZERO,
         )
-
         obj.parent = tst.parent
-        check_flags(
+        check_enum(
             obj=obj,
             tst=tst,
         )
