+++ tests/test_parse_field_list.py
@@ -2,15 +2,14 @@ from __future__ import annotations
 
 import struct
 from collections import OrderedDict
-from typing import Any
+from typing import Any, Sequence
 
-import pytest
 from bitarray import bitarray
 from parse_data import ParseData
 
 from easyprotocol.base.parse_field_list import ParseFieldList
-from easyprotocol.base.parse_generic import DEFAULT_ENDIANNESS
-from easyprotocol.base.parse_value_generic import ParseValueGeneric
+from easyprotocol.base.parse_generic import DEFAULT_ENDIANNESS, ParseGeneric
+from easyprotocol.base.parse_generic_value import ParseGenericValue
 from easyprotocol.fields import UInt8Field
 from easyprotocol.fields.unsigned_int import UIntField
 
@@ -68,9 +67,9 @@ def check_ParseFieldList_children(
             f"{obj}: obj.children[key] is not the expected value "
             + f"({obj.children[key]} != expected value: {tst.children[key]})"
         )
-        assert obj.children[key].parent == obj, (
+        assert obj.children[key]._get_parent_generic() == obj, (  # pyright:ignore[reportPrivateUsage]
             f"{obj}: obj.children[key].parent is not the expected value "
-            + f"({obj.children[key].parent} != expected value: {obj})"
+            + f"({obj.children[key]._get_parent_generic()} != expected value: {obj})"  # pyright:ignore[reportPrivateUsage]
         )
 
     for v in tst.children.values():
@@ -137,8 +136,8 @@ class TestParseFieldList:
         f2 = UInt8Field(name=f2_name)
         bits_data = f2.bits + f1.bits
         byte_data = bytes(f2) + bytes(f1)
-        value: list[Any] = [f1.value, f2.value]
-        children_list: list[ParseValueGeneric[Any]] = [f1, f2]
+        value: list[Any] = [f1, f2]
+        children_list: list[ParseGenericValue[Any]] = [f1, f2]
         tst = ParseData(
             name="test",
             value=value,
@@ -151,7 +150,7 @@ class TestParseFieldList:
         )
         obj = ParseFieldList(
             name=tst.name,
-            children=children_list,
+            default=children_list,
         )
         ParseFieldList_tests(
             obj=obj,
@@ -173,7 +172,7 @@ class TestParseFieldList:
         f2 = UInt8Field(name=f2_name)
         bits_data = f2.bits + f1.bits
         byte_data = bytes(f2) + bytes(f1)
-        values: list[Any] = [f1.value, f2.value]
+        values: list[Any] = [f1, f2]
         tst = ParseData(
             name="test",
             value=values,
@@ -186,7 +185,7 @@ class TestParseFieldList:
         )
         obj = ParseFieldList(
             name=tst.name,
-            children=tst.children,
+            default=tst.children,
         )
         ParseFieldList_tests(
             obj=obj,
@@ -202,7 +201,7 @@ class TestParseFieldList:
         f1 = UInt8Field(name=f1_name)
         byte_data = f1_data
         bits_data = f1_bits
-        values = [f1_value]
+        values = [f1]
         tst = ParseData(
             name="test",
             value=values,
@@ -215,7 +214,7 @@ class TestParseFieldList:
         )
         obj = ParseFieldList(
             name=tst.name,
-            children=[f1],
+            default=[f1],
             data=tst.byte_data,
         )
         ParseFieldList_tests(
@@ -229,26 +228,26 @@ class TestParseFieldList:
         init_bits = bitarray(endian="little")
         init_bits.frombytes(init_data)
         f1_name = "f1"
-        f1_value = 170
+        # f1_value = 170
         f1_data = b"\xaa"
         f1_bits = bitarray(endian="little")
         f1_bits.frombytes(f1_data)
         f1 = UInt8Field(name=f1_name)
         f2_name = "f2"
-        f2_value = 187
+        # f2_value = 187
         f2_data = b"\xbb"
         f2_bits = bitarray(endian="little")
         f2_bits.frombytes(f2_data)
         f2 = UInt8Field(name=f2_name)
         f3_name = "f3"
-        f3_value = 204
+        # f3_value = 204
         f3_data = b"\xcc"
         f3_bits = bitarray(endian="little")
         f3_bits.frombytes(f3_data)
         f3 = UInt8Field(name=f3_name)
         byte_data = f1_data + f2_data + f3_data
         bits_data = f1_bits + f2_bits + f3_bits
-        values = [f1_value, f2_value, f3_value]
+        values = [f1, f2, f3]
         tst = ParseData(
             name="test",
             value=values,
@@ -261,7 +260,7 @@ class TestParseFieldList:
         )
         obj = ParseFieldList(
             name=tst.name,
-            children=[f1, f2, f3],
+            default=[f1, f2, f3],
             data=tst.byte_data,
         )
 
@@ -322,7 +321,7 @@ class TestParseFieldList:
         byte_data = b"\x3F\xF0\x03"
         bits_data = bitarray(endian="little")
         bits_data.frombytes(byte_data)
-        values = [f1_value, f2_value, f3_value, f4_value]
+        values = [f1, f2, f3, f4]
         tst = ParseData(
             name="test",
             value=values,
@@ -335,7 +334,7 @@ class TestParseFieldList:
         )
         obj = ParseFieldList(
             name=tst.name,
-            children=[f1, f2, f3, f4],
+            default=[f1, f2, f3, f4],
             data=tst.byte_data,
         )
 
@@ -354,10 +353,10 @@ class TestParseFieldList:
         byte_data1 = b"\x00"
         bits_data1 = bitarray(endian="little")
         bits_data1.frombytes(byte_data1)
-        values1 = [0]
+        values1 = [f1]
         byte_data2 = f1_data
         bits_data2 = f1_bits
-        values2 = [f1_value]
+        values2 = [f1]
         tst = ParseData(
             name="test",
             value=values1,
@@ -370,7 +369,7 @@ class TestParseFieldList:
         )
         obj = ParseFieldList(
             name=tst.name,
-            children=[f1],
+            default=[f1],
         )
         ParseFieldList_tests(
             obj=obj,
@@ -392,19 +391,19 @@ class TestParseFieldList:
         init_bits = bitarray(endian="little")
         init_bits.frombytes(init_data)
         f1_name = "f1"
-        f1_value = 170
+        # f1_value = 170
         f1_data = b"\xaa"
         f1_bits = bitarray(endian="little")
         f1_bits.frombytes(f1_data)
         f1 = UInt8Field(name=f1_name)
         f2_name = "f2"
-        f2_value = 187
+        # f2_value = 187
         f2_data = b"\xbb"
         f2_bits = bitarray(endian="little")
         f2_bits.frombytes(f2_data)
         f2 = UInt8Field(name=f2_name)
         f3_name = "f3"
-        f3_value = 204
+        # f3_value = 204
         f3_data = b"\xcc"
         f3_bits = bitarray(endian="little")
         f3_bits.frombytes(f3_data)
@@ -412,10 +411,10 @@ class TestParseFieldList:
         byte_data1 = b"\x00\x00\x00"
         bits_data1 = bitarray(endian="little")
         bits_data1.frombytes(byte_data1)
-        values1 = [0, 0, 0]
+        values1 = [f1, f2, f3]
         byte_data2 = f1_data + f2_data + f3_data
         bits_data2 = f1_bits + f2_bits + f3_bits
-        values2 = [f1_value, f2_value, f3_value]
+        values2 = [f1, f2, f3]
         tst = ParseData(
             name="test",
             value=values1,
@@ -429,7 +428,7 @@ class TestParseFieldList:
 
         obj = ParseFieldList(
             name=tst.name,
-            children=[f1, f2, f3],
+            default=[f1, f2, f3],
         )
         ParseFieldList_tests(
             obj=obj,
@@ -476,24 +475,21 @@ class TestParseFieldList:
         )
 
     def test_ParseFieldList_set_value(self) -> None:
-        f1_value = 0
-        f1_data = int.to_bytes(f1_value, length=1, byteorder="big", signed=False)
-        f1_bits = bitarray(endian="little")
-        f1_bits.frombytes(f1_data)
         f2_value = 2
         f2_data = int.to_bytes(f2_value, length=1, byteorder="big", signed=False)
         f2_bits = bitarray(endian="little")
         f2_bits.frombytes(f2_data)
-        f3_value = 20
-        f3_data = int.to_bytes(f3_value, length=1, byteorder="big", signed=False)
-        f3_bits = bitarray(endian="little")
-        f3_bits.frombytes(f3_data)
         f1_name = "f1"
-        f1 = UInt8Field(name=f1_name, default=f1_value)
+        f2 = UInt8Field(name=f1_name, default=f2_value)
         bits_data1 = bitarray(endian="little")
         byte_data1 = b""
-        values1: list[Any] = []
-        children1: OrderedDict[str, ParseValueGeneric[Any]] = OrderedDict()
+        bits_data2 = bitarray(endian="little")
+        bits_data2.frombytes(f2_data)
+        byte_data2 = f2_data
+        values1: Sequence[Any] = []
+        values2: Sequence[Any] = [f2]
+        children1: OrderedDict[str, ParseGenericValue[Any]] = OrderedDict()
+        children2: OrderedDict[str, ParseGenericValue[Any]] = OrderedDict({f2.name: f2})
         tst = ParseData(
             name="test",
             value=values1,
@@ -513,94 +509,15 @@ class TestParseFieldList:
             tst=tst,
         )
 
-        with pytest.raises(NotImplementedError):
-            obj.value = [f1]
-
-    # def test_ParseFieldList_set_value(self) -> None:
-    #     f1_value = 0
-    #     f1_data = int.to_bytes(f1_value, length=1, byteorder="big", signed=False)
-    #     f1_bits = bitarray(endian="little")
-    #     f1_bits.frombytes(f1_data)
-    #     f2_value = 2
-    #     f2_data = int.to_bytes(f2_value, length=1, byteorder="big", signed=False)
-    #     f2_bits = bitarray(endian="little")
-    #     f2_bits.frombytes(f2_data)
-    #     f3_value = 20
-    #     f3_data = int.to_bytes(f3_value, length=1, byteorder="big", signed=False)
-    #     f3_bits = bitarray(endian="little")
-    #     f3_bits.frombytes(f3_data)
-    #     f1_name = "f1"
-    #     f2_name = "f2"
-    #     f1 = UInt8Field(name=f1_name, value=f1_value)
-    #     f2 = UInt8Field(name=f2_name, value=f2_value)
-    #     bits_data1 = bitarray(endian="little")
-    #     bits_data2 = f1_bits
-    #     bits_data3 = f2_bits
-    #     bits_data4 = f3_bits
-    #     byte_data1 = b""
-    #     byte_data2 = f1_data
-    #     byte_data3 = f2_data
-    #     byte_data4 = f3_data
-    #     values1: list[Any] = []
-    #     values2: list[Any] = [f1_value]
-    #     values3: list[Any] = [f2_value]
-    #     values4: list[Any] = [f3_value]
-    #     children1: OrderedDict[str, ParseBaseGeneric[Any]] = OrderedDict()
-    #     children2: OrderedDict[str, ParseBaseGeneric[Any]] = OrderedDict({f1.name: f1})
-    #     children3: OrderedDict[str, ParseBaseGeneric[Any]] = OrderedDict({f1.name: f1})
-    #     children4: OrderedDict[str, ParseBaseGeneric[Any]] = OrderedDict({f1.name: f1})
-    #     tst = ParseData(
-    #         name="test",
-    #         value=values1,
-    #         string_format="{}",
-    #         byte_data=byte_data1,
-    #         bits_data=bits_data1,
-    #         parent=None,
-    #         children=children1,
-    #         endian=DEFAULT_ENDIANNESS,
-    #     )
-
-    #     obj = ParseFieldList(
-    #         name=tst.name,
-    #     )
-    #     ParseFieldList_tests(
-    #         obj=obj,
-    #         tst=tst,
-    #     )
-
-    #     obj.value = [f1]
-    #     tst.value = values2
-    #     tst.byte_data = byte_data2
-    #     tst.bits_data = bits_data2
-    #     tst.children = children2
-    #     ParseFieldList_tests(
-    #         obj=obj,
-    #         tst=tst,
-    #     )
-
-    #     obj.value = [f2]
-    #     tst.value = values3
-    #     tst.byte_data = byte_data3
-    #     tst.bits_data = bits_data3
-    #     tst.children = children3
-    #     ParseFieldList_tests(
-    #         obj=obj,
-    #         tst=tst,
-    #     )
-
-    #     obj.value = [f3_value]
-    #     tst.value = values4
-    #     tst.byte_data = byte_data4
-    #     tst.bits_data = bits_data4
-    #     tst.children = children4
-    #     ParseFieldList_tests(
-    #         obj=obj,
-    #         tst=tst,
-    #     )
-
-    #     value = "invalid"
-    #     with pytest.raises(ValueError):
-    #         obj.value = value  # pyright:ignore[reportGeneralTypeIssues]
+        obj.value = [f2]
+        tst.value = values2
+        tst.bits_data = bits_data2
+        tst.byte_data = byte_data2
+        tst.children = children2
+        ParseFieldList_tests(
+            obj=obj,
+            tst=tst,
+        )
 
     def test_ParseFieldList_remove(self) -> None:
         name = "test"
@@ -608,7 +525,7 @@ class TestParseFieldList:
         f1 = UInt8Field(name=f1_name)
         f2_name = "f2"
         f2 = UInt8Field(name=f2_name)
-        obj = ParseFieldList(name=name, children=[f1, f2])
+        obj = ParseFieldList(name=name, default=[f1, f2])
 
         assert len(obj) == 2
         assert f1.parent == obj
@@ -626,18 +543,21 @@ class TestParseFieldList:
         f3_name = "f3"
         f3 = UInt8Field(name=f3_name, default=3)
         f3_also = UInt8Field(name=f3_name, default=17)
-        obj = ParseFieldList(name=name, children=[f1, f2, f3])
+        obj = ParseFieldList(name=name, default=[f1, f2, f3])
+        i: int = 2
 
         assert f1.parent == obj
         assert f2.parent == obj
         assert f3.parent == obj
         assert f3_also.parent is None
         assert len(obj) == 3
-        assert obj[2] == f3.value
-        obj[2] = f3_also
+        assert obj[2] == f3
+        assert obj[2].value == f3.value
+        obj[i] = f3_also
         assert f3_also.parent == obj
         assert len(obj) == 3
-        assert obj[2] == f3_also.value
+        assert obj[2] == f3_also
+        assert obj[2].value == f3_also.value
 
     def test_ParseFieldList_insert(self) -> None:
         name = "test"
@@ -647,17 +567,22 @@ class TestParseFieldList:
         f2 = UInt8Field(name=f2_name, default=2)
         f3_name = "f3"
         f3 = UInt8Field(name=f3_name, default=3)
-        obj = ParseFieldList(name=name, children=[f1, f3])
+        obj = ParseFieldList(name=name, default=[f1, f3])
 
         assert f1.parent == obj
         assert f2.parent is None
         assert f3.parent == obj
         assert len(obj) == 2
-        assert obj[0] == f1.value
-        assert obj[1] == f3.value
+        assert obj[0] == f1
+        assert obj[0].value == f1.value
+        assert obj[1] == f3
+        assert obj[1].value == f3.value
         obj.insert(1, f2)
         assert f2.parent == obj
         assert len(obj) == 3
-        assert obj[0] == f1.value
-        assert obj[1] == f2.value
-        assert obj[2] == f3.value
+        assert obj[0] == f1
+        assert obj[0].value == f1.value
+        assert obj[1] == f2
+        assert obj[1].value == f2.value
+        assert obj[2] == f3
+        assert obj[2].value == f3.value
